move home dynamic route
